Move StartButton template to Generic
Both brands ship identical start-button markup; brand prefix is handled by the
existing BrandDerivable mixin. Generic now owns the constructor and template,
with Govuk and Nhsuk as empty subclasses — consistent with the rest of the
pilot.
