+++ app/components/design_system/generic/start_button_component.html.erb
